--- src/commons/Constants/Constants.ts
@@ -1,1 +1,0 @@
-export * from './BaseConstants';
