--- src/commons/Constants/Constants.web.ts
@@ -1,18 +1,0 @@
-export * from './BaseConstants';
-import {Constants as BaseConstants} from './BaseConstants';
-
-const WebDeviceFrameSelector = '[data-react-native-web-dimensions="true"]';
-
-function queryWebFrameDevice() {
-  return document?.querySelector(WebDeviceFrameSelector);
-}
-
-export const Constants = {
-  ...BaseConstants,
-  get screenWidth() {
-    return queryWebFrameDevice()?.clientWidth;
-  },
-  get screenHeight() {
-    return queryWebFrameDevice()?.clientHeight;
-  }
-};
