--- src/commons/Constants/index.ts
@@ -1,4 +1,0 @@
-import {Constants} from './Constants';
-
-export * from './Constants';
-export default Constants;
